added dynamic loading of levels and level resource manager
Added `ResourceManager` for static storage of levels and setting/getting the
current level. Would like to have moved the `SubjectFactory` into this
class, due to the way textures are handled in SFML, this was not
possible. See https://en.sfml-dev.org/forums/index.php?topic=7276.0 for
more details. I did move the ownership of the `SubjectFactory` to the
`Game` instead of `World`.